+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added [man2] Independent option and fix
+- Added [man2] `-Independent` option and fix
+- Added [pwmake] `-DeleteCommentEndOfCommandLine` option
 
 
 ## [3.1.0] - 2023-03-18 Sat